ex.12 completed

+++ ch.03/ex.12/Counter.h
@@ -9,15 +9,15 @@ class Counter
 	struct Pair
 	{
 		unsigned short code;
-		size_t matches;
+		unsigned matches;
 
-		Pair(const unsigned short c = 0, const size_t m = 0) : code(c), matches(m) {}
+		Pair(const unsigned short c = 0, const unsigned m = 0) : code(c), matches(m) {}
 		bool operator<(const Pair &other) const;
 	};
 
-	const size_t OFFSET = 32;
+	const unsigned short OFFSET = 32;
 	const unsigned short ARR_SIZE = 95;
-	size_t *arr_code;
+	unsigned *arr_code;
 	std::vector<Pair> sorting; 
 	char format;
